Refactor creation of MidiEvents into Chord
When adding a chord, this used to be MidiHandler's responsibility. This way, however, should be more manageable,
but on the other hand, Chord is now dependent on the Midi library. Although, well, it was before too. But the
point of having MidiHandler was to let it take care of all (or at least many of the) Midi related things. We'll
see how this works.