Fix highlighting issue
By creating BooleanChord, Chord does not need to know about it being selected (reasonable). But due to how onUpdateTrack() in FirstFragment works (it clears the selection booleans), the selections have to be cleared on remove/edit/add buttons pressed.

It also proved much simpler to enable mutli-selections, which I already knew but prefered not to do. I might have to think about a different solution for this.