Refactor the chord track into Progression
The purpose of this is mainly to allow saving and loading different progressions: conversion from
List<Chord> to a saveable format (e. g. json) and vice versa may occur in this class. I opted to keep midiTrack
in MidiHandler, as it is only used for allowing playback, and hence not relevant for Progression. I may want to
move time signature into Progression.